+++ contracts/ValidatorSet/modules/Delegation/VestManager.sol
@@ -53,11 +53,8 @@ contract VestManager is Initializable, OwnableUpgradeable {
         IDelegation(delegation).undelegateWithVesting(validator, amount);
     }
 
-    function swapVestedValidator(address oldValidator, address newValidator) external onlyOwner {
-        uint256 amount = IRewardPool(rewardPool).delegationOf(oldValidator, address(this));
-        _fulfillLiquidTokens(msg.sender, amount);
-        IDelegation(delegation).swapVestedValidator(oldValidator, newValidator);
-        _sendLiquidTokens(msg.sender, amount);
+    function swapVestedPositionValidator(address oldValidator, address newValidator) external onlyOwner {
+        IDelegation(delegation).swapVestedPositionValidator(oldValidator, newValidator);
     }
 
     function claimVestedPositionReward(
