+++ docs/ValidatorSet/ValidatorSet.md
@@ -959,13 +959,13 @@ Stakes sent amount with vesting period.
 |---|---|---|
 | durationWeeks | uint256 | Duration of the vesting in weeks. Must be between 1 and 52. |
 
-### swapVestedValidator
+### swapVestedPositionValidator
 
 ```solidity
-function swapVestedValidator(address oldValidator, address newValidator) external nonpayable
+function swapVestedPositionValidator(address oldValidator, address newValidator) external nonpayable
 ```
 
-Swaps the validator for a vesting position. Can be called by vesting positions&#39; managers only.
+Move a vested position to another validator. Can be called by vesting positions&#39; managers only.
